+++ src/webapp-builtins/tests/HostRouter.test.js
@@ -17,13 +17,6 @@ describe('constructor', () => {
     })).not.toThrow();
   });
 
-  test('accepts a valid minimal configuration with `ignoreCase`', () => {
-    expect(() => new HostRouter({
-      hosts:      {},
-      ignoreCase: false
-    })).not.toThrow();
-  });
-
   test('accepts a valid configuration with several non-wildcard hosts', () => {
     expect(() => new HostRouter({
       hosts: {
@@ -60,9 +53,8 @@ describe('constructor', () => {
     })).not.toThrow();
   });
 
-  test('does not allow two names that differ only in case when `ignoreCase === true`', () => {
+  test('does not allow two names that differ only in case', () => {
     expect(() => new HostRouter({
-      ignoreCase: true,
       hosts: {
         'Boop.bop': 'app1',
         'booP.bop': 'app2'
@@ -70,30 +62,11 @@ describe('constructor', () => {
     })).toThrow();
 
     expect(() => new HostRouter({
-      ignoreCase: true,
       hosts: {
         '*.ZONK': 'app1',
         '*.ZoNK': 'app2'
       }
     })).toThrow();
-  });
-
-  test('allows two names that differ only in case when `ignoreCase === false`', () => {
-    expect(() => new HostRouter({
-      ignoreCase: false,
-      hosts: {
-        'Boop.bop': 'app1',
-        'booP.bop': 'app2'
-      }
-    })).not.toThrow();
-
-    expect(() => new HostRouter({
-      ignoreCase: false,
-      hosts: {
-        '*.ZONK': 'app1',
-        '*.ZoNK': 'app2'
-      }
-    })).not.toThrow();
   });
 
   test.each`
@@ -268,10 +241,9 @@ describe('_impl_handleRequest()', () => {
     await expectApp(hr, 'zorch.splat', 'mockApp1');
   });
 
-  test('routes to a case-folded DNS name when `ignoreCase === true`', async () => {
+  test('routes to a case-folded DNS name', async () => {
     const hr = await makeInstance(
       {
-        ignoreCase: true,
         hosts: {
           'splat':             'mockApp2',
           'ZORCH.splat':       'mockApp1',
@@ -284,10 +256,9 @@ describe('_impl_handleRequest()', () => {
     await expectApp(hr, 'zorch.SPLAT', 'mockApp1');
   });
 
-  test('routes to a case-folded name and not a wildcard, when `ignoreCase === true`', async () => {
+  test('routes to a matching case-folded DNS name and not a wildcard', async () => {
     const hr = await makeInstance(
       {
-        ignoreCase: true,
         hosts: {
           'splat':       'mockApp2',
           'ZORCH.splat': 'mockApp1',
@@ -299,21 +270,6 @@ describe('_impl_handleRequest()', () => {
     );
 
     await expectApp(hr, 'zorCH.splAT', 'mockApp1');
-  });
-
-  test('routes to a case-matched DNS name when `ignoreCase === false`', async () => {
-    const hr = await makeInstance(
-      {
-        ignoreCase: false,
-        hosts: {
-          'ZORCH.splat': 'mockApp1',
-          'zorch.SPLAT': 'mockApp2'
-        }
-      },
-      { appCount: 2 }
-    );
-
-    await expectApp(hr, 'zorch.SPLAT', 'mockApp2');
   });
 
   test('does not route to an exact-match DNS name as if it were a wildcard', async () => {